requirements.txt
Django==5.2.8
djangorestframework==3.16.0
django-cors-headers==4.3.1
djangorestframework_simplejwt==5.5.1
django-allauth
django-db-file-storage==0.5.6.1
django-redis
redis==5.0.1
openai==2.1.0
google-auth==2.42.0
google-genai==1.46.0
dj-database-url==3.0.1
pillow
celery==5.4.0
django-celery-results==2.5.1
channels==4.2
channels_redis==4.2
httpx==0.28.1
django-anymail==14.0
daphne==4.1.2
tiktoken==0.8.0
whitenoise
hiredis>=2.0.0
gunicorn==21.2.0
uvicorn[standard]==0.27.1
psycopg[binary]==3.2.7
python-dotenv==1.0.1
django-storages==1.14.2
boto3==1.34.34
django-cleanup==8.1.0
structlog==24.1.0
drf-spectacular==0.27.1
gevent==23.9.1
structlog